#220 separate tests into Unit And System Testing folders
